+++ Change log.md
@@ -13,3 +13,14 @@
 **MINOR UPDATE 19 JAN 2023**
 - -- ~~do not use letters just reel numbers~~
 - ++ do not use letters just rational numbers
+
+## VERSION PAPERFLOWER 31 MAY 2024 (BETA)
+#### Fully Changed the code !!
+- ++ Can calculate sum, substraction, multiplication, division, permutation, combination, power
+- ++ All mathematical operation compleated with math library. (Except Division)
+- ++ Results now can be saved in forms of " pdf, txt, csv " formats.
+- ++ Results can be saved to a varible but when exited all deleted.
+- +- 2 rational numbers can be used
+- +- do not use letters just rational numbers
+- -- ~~only can calculate all sum, subtract, multiplication, division.~~
+- -- ~~at division it has round option to 0.xy from 0.xyzqwertu....~~